// experimental/angular_projects/projects/Dominion-dev/src/app/dominion/types/card-type.ts
export type CardType =
  | 'Curse'
  | 'Action'
  | 'Treasure'
  | 'Victory'
  | 'Attack'
  | 'Reaction'
  | 'Duration'
  | 'Ruins'
  | 'Prize'
  | 'Looter'
  | 'Shelter'
  | 'Knights'
  | 'Reserve'
  | 'Traveller'
  | 'Castle'
  | 'Gather'
  | 'EventCards'
  | 'LandmarkCards'
  | 'Spirit'
  | 'Night'
  | 'Fate'
  | 'Doom'
  | 'Heirloom'
  | 'Zombie'
  | 'Boon'
  | 'Hex'
  | 'State';